% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FizzBuzz.R
\name{Fizz.Buzz}
\alias{Fizz.Buzz}
\title{Fizz.Buzz Function.}
\usage{
Fizz.Buzz(x)
}
\arguments{
\item{a}{A real number}

\item{b}{A real number}
}
\value{
Any number evenly divisible by 3, substitute "Fizz". Any number evenly divisible by 5,
substitute “Buzz”. Any number divisible by both, substitute “Fizz Buzz”
}
\description{
Child’s game Fizz Buzz. The goal is to count as high as you can, but for any number
evenly divisible by 3, substitute “Fizz” and any number evenly divisible by 5,
substitute “Buzz”, and if it is divisible by both, substitute “Fizz Buzz”.
So the sequence will look like 1, 2, Fizz, 4, Buzz, Fizz, 7, 8, Fizz, . . .
}
\details{
This function will test Fizz Buzz, based on a Youtube video
}
\examples{
Fizz.Buzz(1:30)
}
