% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flagstaff_Weather.R
\docType{data}
\name{Flagstaff_Weather}
\alias{Flagstaff_Weather}
\title{A time series of weather information at Flagstaff's Pulliam Airport from 1950-2019 in Flagstaff, AZ.}
\format{
a data frame with 25449 observations
\describe{
  \item{Date}{The date of observation as a POSIX date format.}
  \item{RAIN}{Daily rain amount in inches.}
  \item{SNOW}{Daily snowfall amount in inches.}
  \item{TMAX}{Daily maximum temperature in degrees Farhenheit.}
  \item{TMIN}{Daily minimum temperature in degrees Farhenheit.}
}
}
\source{
\url{www.ncdc.noaa.gov/cdo-web/search}
}
\usage{
Flagstaff_Weather
}
\description{
A time series of weather information at Flagstaff's Pulliam Airport from 1950-2019 in Flagstaff, AZ.
}
\keyword{datasets}
